gpgpu: fix compile error
fix ptx_sim compile errror

+++ design/gpgpu/gpgpu-sim/src/cuda-sim/ptx_sim.cc
@@ -371,7 +371,7 @@ static void print_reg( FILE *fp, std::string name, ptx_reg_t value, symbol_table
    case U16_TYPE: fprintf(fp,".u16 %u [0x%04x]\n", value.u16, (unsigned) value.u16 ); break;
    case U32_TYPE: fprintf(fp,".u32 %u [0x%08x]\n", value.u32, (unsigned) value.u32 ); break;
    case U64_TYPE: fprintf(fp,".u64 %llu [0x%llx]\n", value.u64, value.u64 ); break;
-   case F16_TYPE: fprintf(fp,".f16 %f [0x%04x]\n",  value.f16, (unsigned) value.u16 ); break;
+   case F16_TYPE: fprintf(fp,".f16 %f [0x%04x]\n",  (float)value.f16, (unsigned) value.u16 ); break;
    case F32_TYPE: fprintf(fp,".f32 %.15lf [0x%08x]\n",  value.f32, value.u32 ); break;
    case F64_TYPE: fprintf(fp,".f64 %.15le [0x%016llx]\n", value.f64, value.u64 ); break;
    case B8_TYPE:  fprintf(fp,".b8  0x%02x\n",   (unsigned) value.u8 );  break;
